[NCL-8044] Add git backend switching
