+++ repour/server/endpoint/internal_scm.py
@@ -1,10 +1,13 @@
 import logging
 
-import asyncssh
 from prometheus_async.aio import time
 from prometheus_client import Histogram, Summary
 
 from repour.config import config
+from repour.server.endpoint import (
+    internal_scm_gerrit,
+    internal_scm_gitlab,
+)
 
 REQ_TIME = Summary("internal_scm_req_time", "time spent with internal-scm endpoint")
 REQ_HISTOGRAM_TIME = Histogram(
@@ -27,54 +30,13 @@ async def internal_scm(spec, repo_provider):
     """
 
     configuration = await config.get_configuration()
-    configuration = configuration.get("gerrit")
+    git_backend = configuration.get("git_backend")
 
-    readonly_url = configuration.get("read_only_template")
-    readwrite_url = configuration.get("read_write_template")
+    logger.debug("Git backend is " + git_backend)
 
-    command = build_gerrit_command(
-        spec.get("project"),
-        spec.get("parent_project"),
-        spec.get("owner_groups"),
-        spec.get("description"),
-    )
-
-    logger.info("Command to run: " + command)
-
-    async with asyncssh.connect(
-        configuration.get("hostname"),
-        username=configuration.get("username"),
-        known_hosts=None,
-    ) as conn:
-        result = await conn.run(command, check=False)
-        exit_status = result.exit_status
-
-        if exit_status == 0:
-            return {
-                "status": "SUCCESS_CREATED",
-                "readonly_url": readonly_url.format(REPO_NAME=spec.get("project")),
-                "readwrite_url": readwrite_url.format(REPO_NAME=spec.get("project")),
-            }
-        elif exit_status == 1 and "Project already exists" in result.stderr:
-            return {
-                "status": "SUCCESS_ALREADY_EXISTS",
-                "readonly_url": readonly_url.format(REPO_NAME=spec.get("project")),
-                "readwrite_url": readwrite_url.format(REPO_NAME=spec.get("project")),
-            }
-        else:
-            raise Exception("Creation failed! Command log: " + result.stderr)
-
-
-def build_gerrit_command(project, parent_project, owner_groups, description):
-    command = "gerrit create-project '{}.git'".format(project)
-
-    if parent_project:
-        command += " -p '{}'".format(parent_project)
-
-    if description:
-        command += " -d '{}'".format(description)
-
-    for owner in owner_groups:
-        command += " -o '{}'".format(owner)
-
-    return command
+    if git_backend == "gitlab":
+        return await internal_scm_gitlab.internal_scm_gitlab(spec, repo_provider)
+    elif git_backend == "gerrit":
+        return await internal_scm_gerrit.internal_scm_gerrit(spec, repo_provider)
+    else:
+        raise Exception("Unknown git server type: " + git_backend)
