+++ repour/server/server.py
@@ -18,7 +18,6 @@ from repour.server.endpoint import (
     info,
     validation,
     internal_scm,
-    internal_scm_gitlab,
 )
 
 logger = logging.getLogger(__name__)
@@ -68,13 +67,6 @@ async def init(loop, bind, repo_provider, repour_url, adjust_provider):
         repour_url,
     )
 
-    internal_scm_gitlab_source = endpoint.validated_json_endpoint(
-        shutdown_callbacks,
-        validation.internal_scm_gitlab,
-        internal_scm_gitlab.internal_scm_gitlab,
-        repour_url,
-    )
-
     logger.debug("Setting up handlers")
     app.router.add_route("GET", "/", info.handle_request)
     app.router.add_route(
@@ -83,7 +75,6 @@ async def init(loop, bind, repo_provider, repour_url, adjust_provider):
     app.router.add_route("POST", "/clone", clone_source)
     app.router.add_route("POST", "/adjust", adjust_source)
     app.router.add_route("POST", "/internal-scm", internal_scm_source)
-    app.router.add_route("POST", "/internal-scm-gitlab", internal_scm_gitlab_source)
     app.router.add_route("POST", "/cancel/{task_id}", cancel.handle_cancel)
     app.router.add_route("GET", "/metrics", aio.web.server_stats)
     app.router.add_route("GET", "/version", info.handle_version)
